scan/requirements.txt
google-cloud-vision==3.4.4
pdf2image==1.16.3
Pillow==10.0.0
requests==2.31.0
python-dotenv==1.0.0 